+++ compiler-plugin-test/src/test/java/io/ballerina/stdlib/persist/compiler/script/CompilerPluginTest.java
@@ -70,8 +70,9 @@ public class CompilerPluginTest {
                 "\tperiod VARCHAR(191),\n" +
                 "\turgency VARCHAR(191),\n" +
                 "\tquantity INT NOT NULL,\n" +
-                "\titemId INT,\n" +
-                "\tCONSTRAINT FK_MEDICAL_NEED_ITEM_0 FOREIGN KEY(itemId) REFERENCES Item(id) ON DELETE CASCADE,\n" +
+                "\titemId INT UNIQUE,\n" +
+                "\tCONSTRAINT FK_MEDICAL_NEED_ITEM_0 FOREIGN KEY(itemId) REFERENCES Item(id) ON DELETE RESTRICT " +
+                "ON UPDATE RESTRICT,\n" +
                 "\tPRIMARY KEY(needId),\n" +
                 "\tUNIQUE KEY(needId)\n" +
                 ")\tAUTO_INCREMENT = 12;";
@@ -109,9 +110,9 @@ public class CompilerPluginTest {
                 "\tperiod VARCHAR(191),\n" +
                 "\turgency VARCHAR(191),\n" +
                 "\tquantity INT NOT NULL,\n" +
-                "\titemId INT,\n" +
-                "\tCONSTRAINT FK_MEDICAL_NEED_MEDICINE_0 FOREIGN KEY(itemId) REFERENCES Medicine(id) ON " +
-                "DELETE CASCADE\n" +
+                "\titemId INT UNIQUE,\n" +
+                "\tCONSTRAINT FK_MEDICAL_NEED_MEDICINE_0 FOREIGN KEY(itemId) REFERENCES Medicine(id) " +
+                "ON DELETE CASCADE\n" +
                 ");";
         testSqlScript("package_03", fileContent, 0, "");
     }
@@ -133,9 +134,9 @@ public class CompilerPluginTest {
                 "\tperiod VARCHAR(191),\n" +
                 "\turgency VARCHAR(191),\n" +
                 "\tquantity INT NOT NULL,\n" +
-                "\titemId INT,\n" +
+                "\titemId INT UNIQUE,\n" +
                 "\tCONSTRAINT FK_MEDICAL_NEED2_MEDICINE2_0 FOREIGN KEY(itemId) REFERENCES Medicine2(id) " +
-                "ON DELETE CASCADE ON UPDATE CASCADE,\n" +
+                "ON DELETE SET NULL ON UPDATE SET NULL,\n" +
                 "\tPRIMARY KEY(needId)\n" +
                 ")\tAUTO_INCREMENT = 12;";
         testSqlScript("package_04", fileContent, 0, "");
@@ -158,8 +159,8 @@ public class CompilerPluginTest {
                 "\tperiod VARCHAR(191) NOT NULL,\n" +
                 "\turgency VARCHAR(191) NOT NULL,\n" +
                 "\tquantity INT NOT NULL,\n" +
-                "\titemId INT,\n" +
-                "\tCONSTRAINT FK_EMPLOYEE_ITEM_0 FOREIGN KEY(itemId) REFERENCES Item(id) ON DELETE CASCADE,\n" +
+                "\titemId INT UNIQUE,\n" +
+                "\tCONSTRAINT FK_EMPLOYEE_ITEM_0 FOREIGN KEY(itemId) REFERENCES Item(id) ON DELETE NO ACTION,\n" +
                 "\tPRIMARY KEY(needId),\n" +
                 "\tUNIQUE KEY(beneficiaryId)\n" +
                 ");";
@@ -185,9 +186,9 @@ public class CompilerPluginTest {
                 "\tid INT NOT NULL AUTO_INCREMENT,\n" +
                 "\tname VARCHAR(191) NOT NULL,\n" +
                 "\titemId INT,\n" +
-                "\tCONSTRAINT FK_ITEM1_ITEM2_0 FOREIGN KEY(itemId) REFERENCES Item2(id) ON DELETE CASCADE,\n" +
+                "\tCONSTRAINT FK_ITEM1_ITEM2_0 FOREIGN KEY(itemId) REFERENCES Item2(id) ON DELETE SET DEFAULT,\n" +
                 "\titemName VARCHAR(191),\n" +
-                "\tCONSTRAINT FK_ITEM1_ITEM2_1 FOREIGN KEY(itemName) REFERENCES Item2(name) ON DELETE CASCADE,\n" +
+                "\tCONSTRAINT FK_ITEM1_ITEM2_1 FOREIGN KEY(itemName) REFERENCES Item2(name) ON DELETE SET DEFAULT,\n" +
                 "\tPRIMARY KEY(id),\n" +
                 "\tUNIQUE KEY(name)\n" +
                 ")\tAUTO_INCREMENT = 5;\n" +
@@ -208,13 +209,15 @@ public class CompilerPluginTest {
                 "\turgency VARCHAR(191) NOT NULL,\n" +
                 "\tquantity INT NOT NULL,\n" +
                 "\titemId INT,\n" +
-                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM_0 FOREIGN KEY(itemId) REFERENCES Item(id) ON DELETE CASCADE,\n" +
+                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM_0 FOREIGN KEY(itemId) REFERENCES Item(id) ON DELETE SET DEFAULT,\n" +
                 "\tname VARCHAR(191),\n" +
-                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM_1 FOREIGN KEY(name) REFERENCES Item(name) ON DELETE CASCADE,\n" +
+                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM_1 FOREIGN KEY(name) REFERENCES Item(name) ON DELETE SET DEFAULT,\n" +
                 "\titemId1 INT,\n" +
-                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM1_0 FOREIGN KEY(itemId1) REFERENCES Item1(id) ON DELETE CASCADE,\n" +
+                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM1_0 FOREIGN KEY(itemId1) REFERENCES Item1(id) " +
+                "ON DELETE SET DEFAULT,\n" +
                 "\tname1 VARCHAR(191),\n" +
-                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM1_1 FOREIGN KEY(name1) REFERENCES Item1(name) ON DELETE CASCADE,\n" +
+                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM1_1 FOREIGN KEY(name1) REFERENCES Item1(name) " +
+                "ON DELETE SET DEFAULT,\n" +
                 "\tPRIMARY KEY(needId),\n" +
                 "\tUNIQUE KEY(beneficiaryId, urgency)\n" +
                 ");";
@@ -246,10 +249,11 @@ public class CompilerPluginTest {
                 "\tperiod VARCHAR(191) NOT NULL,\n" +
                 "\turgency VARCHAR(191) NOT NULL,\n" +
                 "\tquantity INT NOT NULL,\n" +
-                "\titemId INT,\n" +
-                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM_0 FOREIGN KEY(itemId) REFERENCES Item(id) ON DELETE CASCADE,\n" +
-                "\titemId1 INT,\n" +
-                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM1_0 FOREIGN KEY(itemId1) REFERENCES Item1(id) ON DELETE CASCADE,\n" +
+                "\titemId INT UNIQUE,\n" +
+                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM_0 FOREIGN KEY(itemId) REFERENCES Item(id) ON DELETE SET DEFAULT,\n" +
+                "\titemId1 INT UNIQUE,\n" +
+                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM1_0 FOREIGN KEY(itemId1) REFERENCES Item1(id) ON DELETE " +
+                "NO ACTION,\n" +
                 "\tPRIMARY KEY(needId),\n" +
                 "\tUNIQUE KEY(beneficiaryId, urgency)\n" +
                 ");";
@@ -271,8 +275,8 @@ public class CompilerPluginTest {
                 "CREATE TABLE Item (\n" +
                 "\tid INT NOT NULL AUTO_INCREMENT,\n" +
                 "\tname VARCHAR(191) NOT NULL,\n" +
-                "\titemId1 INT,\n" +
-                "\tCONSTRAINT FK_ITEM_ITEM1_0 FOREIGN KEY(itemId1) REFERENCES Item1(id) ON DELETE CASCADE,\n" +
+                "\titemId1 INT UNIQUE,\n" +
+                "\tCONSTRAINT FK_ITEM_ITEM1_0 FOREIGN KEY(itemId1) REFERENCES Item1(id) ON DELETE SET DEFAULT,\n" +
                 "\tPRIMARY KEY(id)\n" +
                 ")\tAUTO_INCREMENT = 3;\n" +
                 "\n" +
@@ -282,8 +286,8 @@ public class CompilerPluginTest {
                 "\tperiod VARCHAR(191) NOT NULL,\n" +
                 "\turgency VARCHAR(191) NOT NULL,\n" +
                 "\tquantity INT NOT NULL,\n" +
-                "\titemId INT,\n" +
-                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM_0 FOREIGN KEY(itemId) REFERENCES Item(id) ON DELETE CASCADE,\n" +
+                "\titemId INT UNIQUE,\n" +
+                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM_0 FOREIGN KEY(itemId) REFERENCES Item(id) ON DELETE SET DEFAULT,\n" +
                 "\tPRIMARY KEY(needId),\n" +
                 "\tUNIQUE KEY(beneficiaryId, urgency)\n" +
                 ");";
@@ -306,8 +310,8 @@ public class CompilerPluginTest {
                 "CREATE TABLE Item (\n" +
                 "\tid INT NOT NULL AUTO_INCREMENT,\n" +
                 "\tname VARCHAR(191) NOT NULL,\n" +
-                "\titemId1 INT,\n" +
-                "\tCONSTRAINT FK_ITEM_ITEM1_0 FOREIGN KEY(itemId1) REFERENCES Item1(id) ON DELETE CASCADE,\n" +
+                "\titemId1 INT UNIQUE,\n" +
+                "\tCONSTRAINT FK_ITEM_ITEM1_0 FOREIGN KEY(itemId1) REFERENCES Item1(id) ON DELETE SET DEFAULT,\n" +
                 "\tPRIMARY KEY(id)\n" +
                 ")\tAUTO_INCREMENT = 3;\n" +
                 "\n" +
@@ -317,8 +321,8 @@ public class CompilerPluginTest {
                 "\tperiod VARCHAR(191) NOT NULL,\n" +
                 "\turgency VARCHAR(191) NOT NULL,\n" +
                 "\tquantity INT NOT NULL,\n" +
-                "\titemId INT,\n" +
-                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM_0 FOREIGN KEY(itemId) REFERENCES Item(id) ON DELETE CASCADE,\n" +
+                "\titemId INT UNIQUE,\n" +
+                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM_0 FOREIGN KEY(itemId) REFERENCES Item(id) ON DELETE SET DEFAULT,\n" +
                 "\tPRIMARY KEY(needId),\n" +
                 "\tUNIQUE KEY(beneficiaryId, urgency)\n" +
                 ");";
@@ -328,15 +332,6 @@ public class CompilerPluginTest {
     @Test
     public void testGenerateSqlScript9() throws IOException {
         String fileContent = "DROP TABLE IF EXISTS MedicalItem;\n" +
-                "CREATE TABLE MedicalItem (\n" +
-                "\tneedId INT NOT NULL AUTO_INCREMENT,\n" +
-                "\titemId INT NOT NULL,\n" +
-                "\tbeneficiaryId INT NOT NULL,\n" +
-                "\tperiod VARCHAR(191) NOT NULL,\n" +
-                "\turgency VARCHAR(191) NOT NULL,\n" +
-                "\tquantity INT NOT NULL,\n" +
-                "\tPRIMARY KEY(needId)\n" +
-                ");\n" +
                 "\n" +
                 "DROP TABLE IF EXISTS MedicalNeeds;\n" +
                 "\n" +
@@ -352,8 +347,8 @@ public class CompilerPluginTest {
                 "CREATE TABLE Item (\n" +
                 "\tid INT NOT NULL AUTO_INCREMENT,\n" +
                 "\tname VARCHAR(191) NOT NULL,\n" +
-                "\titemId1 INT,\n" +
-                "\tCONSTRAINT FK_ITEM_ITEM1_0 FOREIGN KEY(itemId1) REFERENCES Item1(id) ON DELETE CASCADE,\n" +
+                "\titemId1 INT UNIQUE,\n" +
+                "\tCONSTRAINT FK_ITEM_ITEM1_0 FOREIGN KEY(itemId1) REFERENCES Item1(id) ON DELETE SET DEFAULT,\n" +
                 "\tPRIMARY KEY(id)\n" +
                 ")\tAUTO_INCREMENT = 3;\n" +
                 "\n" +
@@ -363,10 +358,20 @@ public class CompilerPluginTest {
                 "\tperiod VARCHAR(191) NOT NULL,\n" +
                 "\turgency VARCHAR(191) NOT NULL,\n" +
                 "\tquantity INT NOT NULL,\n" +
-                "\titemId INT,\n" +
-                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM_0 FOREIGN KEY(itemId) REFERENCES Item(id) ON DELETE CASCADE,\n" +
+                "\titemId INT UNIQUE,\n" +
+                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM_0 FOREIGN KEY(itemId) REFERENCES Item(id) ON DELETE SET DEFAULT,\n" +
                 "\tPRIMARY KEY(needId),\n" +
                 "\tUNIQUE KEY(beneficiaryId, urgency)\n" +
+                ");\n" +
+                "\n" +
+                "CREATE TABLE MedicalItem (\n" +
+                "\tneedId INT NOT NULL AUTO_INCREMENT,\n" +
+                "\titemId INT NOT NULL,\n" +
+                "\tbeneficiaryId INT NOT NULL,\n" +
+                "\tperiod VARCHAR(191) NOT NULL,\n" +
+                "\turgency VARCHAR(191) NOT NULL,\n" +
+                "\tquantity INT NOT NULL,\n" +
+                "\tPRIMARY KEY(needId)\n" +
                 ");";
         String fileContent1 = "DROP TABLE IF EXISTS MedicalNeeds;\n" +
                 "\n" +
@@ -382,8 +387,8 @@ public class CompilerPluginTest {
                 "CREATE TABLE Item (\n" +
                 "\tid INT NOT NULL AUTO_INCREMENT,\n" +
                 "\tname VARCHAR(191) NOT NULL,\n" +
-                "\titemId1 INT,\n" +
-                "\tCONSTRAINT FK_ITEM_ITEM1_0 FOREIGN KEY(itemId1) REFERENCES Item1(id) ON DELETE CASCADE,\n" +
+                "\titemId1 INT UNIQUE,\n" +
+                "\tCONSTRAINT FK_ITEM_ITEM1_0 FOREIGN KEY(itemId1) REFERENCES Item1(id) ON DELETE SET DEFAULT,\n" +
                 "\tPRIMARY KEY(id)\n" +
                 ")\tAUTO_INCREMENT = 3;\n" +
                 "\n" +
@@ -393,8 +398,8 @@ public class CompilerPluginTest {
                 "\tperiod VARCHAR(191) NOT NULL,\n" +
                 "\turgency VARCHAR(191) NOT NULL,\n" +
                 "\tquantity INT NOT NULL,\n" +
-                "\titemId INT,\n" +
-                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM_0 FOREIGN KEY(itemId) REFERENCES Item(id) ON DELETE CASCADE,\n" +
+                "\titemId INT UNIQUE,\n" +
+                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM_0 FOREIGN KEY(itemId) REFERENCES Item(id) ON DELETE SET DEFAULT,\n" +
                 "\tPRIMARY KEY(needId),\n" +
                 "\tUNIQUE KEY(beneficiaryId, urgency)\n" +
                 ");\n" +
@@ -441,9 +446,9 @@ public class CompilerPluginTest {
                 "\tid INT NOT NULL AUTO_INCREMENT,\n" +
                 "\tname VARCHAR(20) NOT NULL,\n" +
                 "\titemId INT,\n" +
-                "\tCONSTRAINT FK_ITEM1_ITEM2_0 FOREIGN KEY(itemId) REFERENCES Item2(id) ON DELETE CASCADE,\n" +
+                "\tCONSTRAINT FK_ITEM1_ITEM2_0 FOREIGN KEY(itemId) REFERENCES Item2(id) ON DELETE SET DEFAULT,\n" +
                 "\titemName VARCHAR(20),\n" +
-                "\tCONSTRAINT FK_ITEM1_ITEM2_1 FOREIGN KEY(itemName) REFERENCES Item2(name) ON DELETE CASCADE,\n" +
+                "\tCONSTRAINT FK_ITEM1_ITEM2_1 FOREIGN KEY(itemName) REFERENCES Item2(name) ON DELETE SET DEFAULT,\n" +
                 "\tPRIMARY KEY(id),\n" +
                 "\tUNIQUE KEY(name)\n" +
                 ")\tAUTO_INCREMENT = 5;\n" +
@@ -464,17 +469,162 @@ public class CompilerPluginTest {
                 "\turgency VARCHAR(191) NOT NULL,\n" +
                 "\tquantity INT NOT NULL,\n" +
                 "\titemId INT,\n" +
-                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM_0 FOREIGN KEY(itemId) REFERENCES Item(id) ON DELETE CASCADE,\n" +
+                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM_0 FOREIGN KEY(itemId) REFERENCES Item(id) ON DELETE SET DEFAULT,\n" +
                 "\tname VARCHAR(10),\n" +
-                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM_1 FOREIGN KEY(name) REFERENCES Item(name) ON DELETE CASCADE,\n" +
+                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM_1 FOREIGN KEY(name) REFERENCES Item(name) ON DELETE SET DEFAULT,\n" +
                 "\titemId1 INT,\n" +
-                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM1_0 FOREIGN KEY(itemId1) REFERENCES Item1(id) ON DELETE CASCADE,\n" +
+                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM1_0 FOREIGN KEY(itemId1) REFERENCES Item1(id) " +
+                "ON DELETE SET DEFAULT,\n" +
                 "\tname1 VARCHAR(20),\n" +
-                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM1_1 FOREIGN KEY(name1) REFERENCES Item1(name) ON DELETE CASCADE,\n" +
+                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM1_1 FOREIGN KEY(name1) REFERENCES Item1(name) " +
+                "ON DELETE SET DEFAULT,\n" +
                 "\tPRIMARY KEY(needId),\n" +
                 "\tUNIQUE KEY(beneficiaryId, urgency)\n" +
                 ");";
         testSqlScript("package_11", fileContent, 0, "");
+    }
+
+    @Test
+    public void testGenerateSqlScript11() throws IOException {
+        String fileContent = "DROP TABLE IF EXISTS MedicalNeeds;\n" +
+                "\n" +
+                "DROP TABLE IF EXISTS Item1;\n" +
+                "\n" +
+                "DROP TABLE IF EXISTS Item2;\n" +
+                "CREATE TABLE Item2 (\n" +
+                "\tid INT NOT NULL AUTO_INCREMENT,\n" +
+                "\tname VARCHAR(20) NOT NULL,\n" +
+                "\tPRIMARY KEY(id)\n" +
+                ")\tAUTO_INCREMENT = 2;\n" +
+                "\n" +
+                "CREATE TABLE Item1 (\n" +
+                "\tid INT NOT NULL AUTO_INCREMENT,\n" +
+                "\tname VARCHAR(20) NOT NULL,\n" +
+                "\titemName VARCHAR(20),\n" +
+                "\tCONSTRAINT FK_ITEM1_ITEM2_0 FOREIGN KEY(itemName) REFERENCES Item2(name) ON DELETE SET DEFAULT,\n" +
+                "\tPRIMARY KEY(id),\n" +
+                "\tUNIQUE KEY(name)\n" +
+                ")\tAUTO_INCREMENT = 5;\n" +
+                "\n" +
+                "\n" +
+                "DROP TABLE IF EXISTS Item;\n" +
+                "CREATE TABLE Item (\n" +
+                "\tid INT NOT NULL AUTO_INCREMENT,\n" +
+                "\tname VARCHAR(10) NOT NULL,\n" +
+                "\tPRIMARY KEY(id),\n" +
+                "\tUNIQUE KEY(name)\n" +
+                ")\tAUTO_INCREMENT = 3;\n" +
+                "\n" +
+                "CREATE TABLE MedicalNeeds (\n" +
+                "\tneedId INT NOT NULL AUTO_INCREMENT,\n" +
+                "\tbeneficiaryId INT NOT NULL,\n" +
+                "\tperiod VARCHAR(191) NOT NULL,\n" +
+                "\turgency VARCHAR(191) NOT NULL,\n" +
+                "\tquantity INT NOT NULL,\n" +
+                "\tname VARCHAR(10) UNIQUE,\n" +
+                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM_0 FOREIGN KEY(name) REFERENCES Item(name),\n" +
+                "\tname1 VARCHAR(20) UNIQUE,\n" +
+                "\tCONSTRAINT FK_MEDICALNEEDS_ITEM1_0 FOREIGN KEY(name1) REFERENCES Item1(name),\n" +
+                "\tPRIMARY KEY(needId),\n" +
+                "\tUNIQUE KEY(beneficiaryId, urgency)\n" +
+                ");";
+        testSqlScript("package_12", fileContent, 0, "");
+    }
+
+    @Test
+    public void testGenerateSqlScript12() throws IOException {
+        String fileContent = "DROP TABLE IF EXISTS Profile;\n" +
+                "\n" +
+                "DROP TABLE IF EXISTS User;\n" +
+                "CREATE TABLE User (\n" +
+                "\tid INT NOT NULL,\n" +
+                "\tname VARCHAR(191) NOT NULL,\n" +
+                "\tPRIMARY KEY(id)\n" +
+                ");\n" +
+                "\n" +
+                "CREATE TABLE Profile (\n" +
+                "\tid INT NOT NULL,\n" +
+                "\tname VARCHAR(191) NOT NULL,\n" +
+                "\tuserId INT UNIQUE,\n" +
+                "\tCONSTRAINT FK_PROFILE_USER_0 FOREIGN KEY(userId) REFERENCES User(id),\n" +
+                "\tPRIMARY KEY(id)\n" +
+                ");";
+        testSqlScript("package_13", fileContent, 0, "");
+    }
+
+    @Test
+    public void testGenerateSqlScript13() throws IOException {
+        String fileContent = "DROP TABLE IF EXISTS Post;\n" +
+                "\n" +
+                "DROP TABLE IF EXISTS User;\n" +
+                "CREATE TABLE User (\n" +
+                "\tid INT NOT NULL,\n" +
+                "\tname VARCHAR(191) NOT NULL,\n" +
+                "\tPRIMARY KEY(id)\n" +
+                ");\n" +
+                "\n" +
+                "CREATE TABLE Post (\n" +
+                "\tid INT NOT NULL,\n" +
+                "\tname VARCHAR(191) NOT NULL,\n" +
+                "\tauthorId INT,\n" +
+                "\tCONSTRAINT FK_POST_USER_0 FOREIGN KEY(authorId) REFERENCES User(id),\n" +
+                "\tPRIMARY KEY(id)\n" +
+                ");";
+        testSqlScript("package_14", fileContent, 0, "");
+    }
+
+    @Test
+    public void testGenerateSqlScript14() throws IOException {
+        String fileContent = "DROP TABLE IF EXISTS POST_TABLE;\n" +
+                "\n" +
+                "DROP TABLE IF EXISTS USER_TABLE;\n" +
+                "CREATE TABLE USER_TABLE (\n" +
+                "\tid INT NOT NULL,\n" +
+                "\tname VARCHAR(191) NOT NULL,\n" +
+                "\tPRIMARY KEY(id)\n" +
+                ");\n" +
+                "\n" +
+                "CREATE TABLE POST_TABLE (\n" +
+                "\tid INT NOT NULL,\n" +
+                "\tname VARCHAR(191) NOT NULL,\n" +
+                "\tauthorId INT,\n" +
+                "\tCONSTRAINT FK_POST_TABLE_USER_TABLE_0 FOREIGN KEY(authorId) REFERENCES USER_TABLE(id),\n" +
+                "\tPRIMARY KEY(id)\n" +
+                ");";
+        testSqlScript("package_15", fileContent, 0, "");
+    }
+
+    @Test
+    public void testGenerateSqlScript15() throws IOException {
+        String fileContent = "DROP TABLE IF EXISTS Multiple_Associations;\n" +
+                "\n" +
+                "DROP TABLE IF EXISTS Profiles;\n" +
+                "\n" +
+                "DROP TABLE IF EXISTS Users;\n" +
+                "CREATE TABLE Users (\n" +
+                "\tid INT NOT NULL,\n" +
+                "\tname VARCHAR(191) NOT NULL,\n" +
+                "\tPRIMARY KEY(id)\n" +
+                ");\n" +
+                "\n" +
+                "CREATE TABLE Profiles (\n" +
+                "\tid INT NOT NULL,\n" +
+                "\tname VARCHAR(191) NOT NULL,\n" +
+                "\tuserId INT UNIQUE,\n" +
+                "\tCONSTRAINT FK_PROFILES_USERS_0 FOREIGN KEY(userId) REFERENCES Users(id),\n" +
+                "\tPRIMARY KEY(id)\n" +
+                ");\n" +
+                "\n" +
+                "CREATE TABLE Multiple_Associations (\n" +
+                "\tid INT NOT NULL,\n" +
+                "\tname VARCHAR(191) NOT NULL,\n" +
+                "\tprofileId INT UNIQUE,\n" +
+                "\tCONSTRAINT FK_MULTIPLE_ASSOCIATIONS_PROFILES_0 FOREIGN KEY(profileId) REFERENCES Profiles(id),\n" +
+                "\tuserId INT UNIQUE,\n" +
+                "\tCONSTRAINT FK_MULTIPLE_ASSOCIATIONS_USERS_0 FOREIGN KEY(userId) REFERENCES Users(id),\n" +
+                "\tPRIMARY KEY(id)\n" +
+                ");";
+        testSqlScript("package_16", fileContent, 0, "");
     }
 
     private void testSqlScript(String packagePath, String fileContent, int count,