+++ compiler-plugin/src/main/java/io/ballerina/stdlib/persist/compiler/Constants.java
@@ -18,6 +18,13 @@
 
 package io.ballerina.stdlib.persist.compiler;
 
+import io.ballerina.compiler.syntax.tree.LiteralValueToken;
+import io.ballerina.compiler.syntax.tree.NodeFactory;
+import io.ballerina.compiler.syntax.tree.SyntaxKind;
+import io.ballerina.compiler.syntax.tree.Token;
+
+import static io.ballerina.compiler.syntax.tree.AbstractNodeFactory.createEmptyMinutiaeList;
+
 /**
  * Constants for Persist compiler plugin.
  */
@@ -25,7 +32,7 @@ public class Constants {
 
     public static final String TRUE = "true";
     public static final String INSERT_METHOD_NAME = "runInsertQuery";
-    public static final String ENTITY = "persist:Entity ";
+    public static final String ENTITY = "persist:Entity";
     public static final String UNNECESSARY_CHARS_REGEX = "\"|\\n";
     public static final String TABLE_NAME = "tableName";
     public static final String AUTO_INCREMENT = "persist:AutoIncrement";
@@ -35,9 +42,6 @@ public class Constants {
     public static final String REFERENCE = "reference";
     public static final String START_VALUE = "startValue";
     public static final String ONE = "1";
-    public static final String CASCADE_DELETE = "cascadeDelete";
-    public static final String ON_DELETE_CASCADE = " ON DELETE CASCADE";
-    public static final String ON_UPDATE_CASCADE = " ON UPDATE CASCADE";
     public static final String NOT_NULL = " NOT NULL";
     public static final String AUTO_INCREMENT_WITH_SPACE = " AUTO_INCREMENT";
     public static final String AUTO_INCREMENT_WITH_TAB = "\tAUTO_INCREMENT";
@@ -46,6 +50,19 @@ public class Constants {
     public static final String MAX_LENGTH = "maxLength";
     public static final String FILE_NAME = "persist_db_scripts.sql";
     public static final String VARCHAR_LENGTH = "191";
+    public static final String ONE_TO_ONE = "ONE";
+    public static final String ONE_TO_MANY = "MANY";
+    public static final String KEY = "key";
+    public static final String UNIQUE_CONSTRAINTS = "uniqueConstraints";
+
+    public static final String READ_FUNCTION = "read";
+    public static final String EXECUTE_FUNCTION = "execute";
+    public static final String BACKTICK = "`";
+    public static final String SPACE = " ";
+    public static final String OPEN_BRACES = "( ";
+    public static final String CLOSE_BRACES = ") ";
+    public static final String ASCENDING = "ascending";
+    public static final String BAL_ESCAPE_TOKEN = "'";
 
     /**
      * Constants related to Ballerina types.
@@ -76,4 +93,31 @@ public class Constants {
         public static final String TIME_STAMP = "TIMESTAMP";
         public static final String DATE_TIME = "DATETIME";
     }
+
+    /**
+     * SQL keywords used to construct the query.
+     */
+    public static final class SQLKeyWords {
+        public static final String WHERE = "WHERE";
+        public static final String LIMIT = "LIMIT";
+        public static final String ORDERBY = "ORDER BY";
+        public static final String ORDER_BY_ASCENDING = "ASC";
+        public static final String ORDER_BY_DECENDING = "DESC";
+        public static final String NOT_EQUAL_TOKEN = "<>";
+        public static final String AND = "AND";
+        public static final String OR = "OR";
+    }
+
+    /**
+     * Constant nodes used in code modification.
+     */
+    public static final class TokenNodes {
+        public static final Token INTERPOLATION_START_TOKEN = NodeFactory.createLiteralValueToken(
+                SyntaxKind.INTERPOLATION_START_TOKEN, "${", createEmptyMinutiaeList(), createEmptyMinutiaeList());
+        public static final Token INTERPOLATION_END_TOKEN = NodeFactory.createLiteralValueToken(
+                SyntaxKind.CLOSE_BRACE_TOKEN, "}", createEmptyMinutiaeList(), createEmptyMinutiaeList());
+        public static final LiteralValueToken BACKTICK_TOKEN = NodeFactory.createLiteralValueToken(
+                SyntaxKind.BACKTICK_TOKEN, BACKTICK, createEmptyMinutiaeList(), createEmptyMinutiaeList());
+    }
+
 }