+++ compiler-plugin/src/main/java/io/ballerina/stdlib/persist/compiler/DiagnosticsCodes.java
@@ -51,7 +51,15 @@ public enum DiagnosticsCodes {
     PERSIST_112("PERSIST_112", "mysql db only allow increment value by one in auto generated field",
             WARNING),
     PERSIST_113("PERSIST_113", "duplicate table name: the table name is already used in another entity",
-            ERROR);
+            ERROR),
+    PERSIST_114("PERSIST_114", "unsupported features: many-to-many association is not supported yet",
+            ERROR),
+    PERSIST_115("PERSIST_115", "invalid entity initialisation: the associated entity of this[{0}] " +
+            "does not have the field with the relationship type", ERROR),
+    PERSIST_116("PERSIST_116", "invalid entity initialisation: the relation annotation should " +
+            "only be added to the relationship owner for one-to-one and one-to-many associations", ERROR),
+    PERSIST_117("PERSIST_117", "invalid annotation attachment: The relation annotation " +
+            "can only be attached to entity typed fields", ERROR);
 
     private final String code;
     private final String message;